requirements.txt
Django              4.1.5
django-allauth      0.52.0
djangorestframework 3.14.0